chore: updates for BuildLimit int -> int64

+++ action/repo/repo.go
@@ -14,7 +14,7 @@ type Config struct {
 	Link         string
 	Clone        string
 	Visibility   string
-	BuildLimit   int
+	BuildLimit   int64
 	Timeout      int64
 	Counter      int
 	Private      bool
